Feat: cleaned the map code for better understanding and scalablity
